backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pandas>=2.2.0
groq>=0.5.0
sqlalchemy>=2.0.0
pytest>=8.0.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
